feat: add routing, polish card screens
