+++ src/widgets/card/style.ts
@@ -1,0 +1,11 @@
+import { SxProps } from '@mui/material';
+
+export const buttonStyle: SxProps = {
+  fontSize: '0.875rem',
+  padding: '1.125rem 0',
+  textTransform: 'none',
+};
+export const topButtonsStyle: SxProps = {
+  paddingTop: 2.5,
+  paddingBottom: 1.5,
+};
