+++ src/widgets/index.ts
@@ -3,3 +3,4 @@ export * from './sign-in';
 export * from './sign-up';
 export * from './authorization';
 export * from './cards-list';
+export * from './card';
